Speed up training on large systems 50% by replacing jvp with linearize.
In the kinetic energy evaluation, the primal computation of `grad_f_closure` was being reevaluated every iteration of the loop to compute the Laplacian. By using jax.linearize, the primal only has to be evaluated once, roughly doubling the speed of the kinetic energy evaluation and increasing training speed 50% on large systems.

PiperOrigin-RevId: 434449652
Change-Id: Ie0996d37b5e7541e6b87e53bca933b8b697bee4d